refactor & cleanup
